Update the fucking license
OK, I give up. This should be the final license before I start the
school. If you found any error, fork this and edit it yourself, I don't
give a fuck anymore.

I have life, school and the future. It is not yours, so don't give a
fuck about it